+++ src/Table/Table.story.js
@@ -1,12 +1,35 @@
+import React from 'react';
+import Paper from '@material-ui/core/Paper';
 import { storiesOf } from '@storybook/react';
+import Table from './Table';
+import { cars } from './exampleData';
 import { GROUPS } from '../../.storybook/hierarchySeparators';
-import SimpleTable from './SimpleTable';
-import FilteringTable from './FilteredTable';
-import SelectionTable from './SelectionTable';
-import ActionTable from './ActionTable';
 
-storiesOf(`${GROUPS.COMPONENTS}|Table`, module)
-  .add('Simple Table', SimpleTable)
-  .add('Filtering Table', FilteringTable)
-  .add('Selection Table', SelectionTable)
-  .add('Action Table', ActionTable);
+const columns = [
+  {
+    title: 'Brand',
+    field: 'brand',
+  },
+  {
+    title: 'Model',
+    field: 'model.name',
+  },
+  {
+    title: 'Year',
+    field: 'model.year',
+    options: {
+      numeric: true,
+    },
+  },
+];
+
+const SimpleTable = () => (
+  <Paper style={{ width: 1000 }}>
+    <Table columns={columns} data={cars} rowId={row => row.id} />
+  </Paper>
+);
+
+storiesOf(`${GROUPS.COMPONENTS}|Table`, module).add(
+  'Simple Table',
+  SimpleTable
+);
