+++ src/Table/TableAction.story.js
@@ -4,8 +4,10 @@ import EditIcon from '@material-ui/icons/Edit';
 import SearchIcon from '@material-ui/icons/Search';
 import MailIcon from '@material-ui/icons/Mail';
 import DeleteIcon from '@material-ui/icons/DeleteForever';
+import { storiesOf } from '@storybook/react';
 import Table from './Table';
 import { cars } from './exampleData';
+import { GROUPS } from '../../.storybook/hierarchySeparators';
 
 const columns = [
   {
@@ -73,4 +75,7 @@ const ActionTable = () => (
   </Paper>
 );
 
-export default ActionTable;
+storiesOf(`${GROUPS.COMPONENTS}|Table`, module).add(
+  'Action Table',
+  ActionTable
+);
