+++ src/Table/TableFiltered.story.js
@@ -1,7 +1,9 @@
 import React from 'react';
 import Paper from '@material-ui/core/Paper';
+import { storiesOf } from '@storybook/react';
 import Table from './Table';
 import { cars } from './exampleData';
+import { GROUPS } from '../../.storybook/hierarchySeparators';
 
 const columns = [
   {
@@ -44,4 +46,7 @@ const FilteringTable = () => (
   </Paper>
 );
 
-export default FilteringTable;
+storiesOf(`${GROUPS.COMPONENTS}|Table`, module).add(
+  'Filtering Table',
+  FilteringTable
+);
